# asg_form/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ASG 赛事官网-后端API文档",
    "description": "这是由罗澜使用ASP.NET.Core开发的ASG赛事组后端系统，包括官网和后台管理系统。使用 sqlserver作为数据库，identity框架进行账号控制。",
    "version": "V 1.9.6"
  },
  "paths": {
...
    "/api/v1/admin/allperson": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "获取所有用户-支持分页",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page_long",
            "description": "",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/post_user"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/deluser": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "删除用户,需要superadmin",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "description": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/admin/enroll": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "管理员直接添加一个用户",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "captoken",
            "description": "",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/newuser_get"
            }
          }
        }
      }
    },
    "/api/v1/admin/form": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "删除表单",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "formname",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/v1/admin/officium": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "设置职位,需要superadmin",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "description": "用户id",
            "type": "string"
          },
          {
            "in": "query",
            "name": "officium",
            "description": "职位名称",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/admin/setadmin": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "设置管理员,需要superadmin",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "description": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "newuser_get": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "chinaname": {
          "type": "string"
        },
        "eMail": {
          "type": "string"
        }
      }
    },
...
    "post_user": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "chinaname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "isadmin": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}